Russekort/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
qrcode==7.4.2
Pillow==10.4.0
python-multipart==0.0.9

